[OvsPpd] Changed to uint32_t probability
Seemed pretty much necessary, given that the kernel part of the datapath
hates float usage with a passion. In retrospect, I probably should have
seen this coming from the start.

+++ code/ovs-ppd/include/openvswitch/ofp-actions.h
@@ -1019,7 +1019,7 @@ struct ofpact_decap {
  * Used for OFPAT_PROBDROP */
 struct ofpact_probdrop {
     struct ofpact ofpact;
-    float prob;           /* Float probability, punned to int on receipt. */
+    uint32_t prob;           /* Uint probability, "covers" 0->1 range. */
 };
 
 /* Converting OpenFlow to ofpacts. */
